Queue of functions
Added a FunctionQueue in the main.cpp (should probably be moved to TestFunctor but it was easier to test in Main). The FunctionQueue is a list of objects that can be functions, function pointers, functors, or lambdas. Created two "test" functions hello and divideByZero and passed it in to the FunctionQueue then executed from the queue. Also included many comments lol perhaps too many